refactor: 토픽&키워드 관련 명령-조회 분리 (#97)
